+++ src/main/java/de/samply/csv/ContainerCsvWriterIterable.java
@@ -1,6 +1,7 @@
 package de.samply.csv;
 
 import de.samply.container.Container;
+import de.samply.files.ContainerFileWriterIterable;
 import de.samply.template.AttributeTemplate;
 import de.samply.template.ContainerTemplate;
 import de.samply.template.ConverterTemplate;
@@ -8,34 +9,23 @@ import java.util.Iterator;
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
 
-public class ContainerCsvWriterIterable implements Iterable<String> {
+public class ContainerCsvWriterIterable extends ContainerFileWriterIterable {
 
-  private List<Container> containers;
-  private ContainerTemplate containerTemplate;
-  private ConverterTemplate converterTemplate;
-  private boolean headersExists;
+  private boolean headersExists = false;
 
-  public ContainerCsvWriterIterable(List<Container> containers, ConverterTemplate converterTemplate, ContainerTemplate containerTemplate, boolean headersExists) {
-    this.containers = containers;
-    this.converterTemplate = converterTemplate;
-    this.containerTemplate = containerTemplate;
-    this.headersExists = headersExists;
+  public ContainerCsvWriterIterable(List<Container> containers, ConverterTemplate converterTemplate,
+      ContainerTemplate containerTemplate) {
+    super(containers, converterTemplate, containerTemplate);
   }
 
   @NotNull
   @Override
   public Iterator<String> iterator() {
-    return new ContainerWriterIterator();
+    return new ContainerCsvWriterIterator();
   }
 
-  private class ContainerWriterIterator implements Iterator<String>{
-
+  private class ContainerCsvWriterIterator extends ContainerWriterIterator{
     private boolean isHeaderLine = !headersExists;
-    private Iterator<Container> containerIterator;
-
-    public ContainerWriterIterator() {
-      this.containerIterator = containers.iterator();
-    }
 
     @Override
     public boolean hasNext() {
@@ -44,38 +34,48 @@ public class ContainerCsvWriterIterable implements Iterable<String> {
 
     @Override
     public String next() {
-      String line = (isHeaderLine) ? fetchHeaderLine() : fetchAttributeLine();
+      String line = (isHeaderLine) ? fetchHeaderLine() : fetchAttributeLine(containerIterator.next());
       isHeaderLine = false;
       return line;
     }
 
-    private String fetchHeaderLine(){
+    private String fetchHeaderLine() {
       List<AttributeTemplate> templates = containerTemplate.getAttributeTemplates();
       StringBuilder stringBuilder = new StringBuilder();
       templates.forEach(attributeTemplate -> {
         stringBuilder.append(attributeTemplate.getCsvColumnName());
         stringBuilder.append(converterTemplate.getCsvSeparator());
       });
-      stringBuilder.deleteCharAt(stringBuilder.length()-1);
+      stringBuilder.deleteCharAt(stringBuilder.length() - 1);
       return stringBuilder.toString();
     }
 
-    private String fetchAttributeLine(){
-      Container container = containerIterator.next();
+    private String fetchAttributeLine(Container container) {
       List<AttributeTemplate> templates = containerTemplate.getAttributeTemplates();
       StringBuilder stringBuilder = new StringBuilder();
       templates.forEach(attributeTemplate -> {
         String attributeValue = container.getAttributeValue(attributeTemplate);
-        if (attributeValue == null){
+        if (attributeValue == null) {
           attributeValue = "";
         }
         stringBuilder.append(attributeValue);
         stringBuilder.append(converterTemplate.getCsvSeparator());
       });
-      stringBuilder.deleteCharAt(stringBuilder.length()-1);
+      stringBuilder.deleteCharAt(stringBuilder.length() - 1);
       return stringBuilder.toString();
     }
 
   }
+
+  @Override
+  protected String fetchLine(Container container) {
+    return null;
+  }
+
+  public ContainerCsvWriterIterable headerExists(boolean headersExists){
+    this.headersExists = headersExists;
+    return this;
+  }
+
 
 }