+++ src/main/java/de/samply/template/ContainerTemplate.java
@@ -10,6 +10,14 @@ public class ContainerTemplate {
   @JacksonXmlProperty(isAttribute = true, localName = "csv-filename")
   @JsonProperty("csv-filename")
   private String csvFilename;
+  @JacksonXmlProperty(isAttribute = true, localName = "json-filename")
+  @JsonProperty("json-filename")
+  private String jsonFilename;
+
+  @JacksonXmlProperty(isAttribute = true, localName = "json-key")
+  @JsonProperty("json-key")
+  private String jsonKey;
+
   @JacksonXmlProperty(isAttribute = true, localName = "excel-sheet")
   @JsonProperty("excel-sheet")
   private String excelSheet;
@@ -72,6 +80,22 @@ public class ContainerTemplate {
   public void setAttributeTemplates(
       List<AttributeTemplate> attributeTemplates) {
     this.attributeTemplates = attributeTemplates;
+  }
+
+  public String getJsonFilename() {
+    return jsonFilename;
+  }
+
+  public void setJsonFilename(String jsonFilename) {
+    this.jsonFilename = jsonFilename;
+  }
+
+  public String getJsonKey() {
+    return jsonKey;
+  }
+
+  public void setJsonKey(String jsonKey) {
+    this.jsonKey = jsonKey;
   }
 
 }